With Sping and IOC
