+++ .idea/compiler.xml
@@ -6,6 +6,7 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="myPersonalChatBot" />
         <module name="io.ayush.bot" />
       </profile>
     </annotationProcessing>
