+++ pom.xml
@@ -35,8 +35,15 @@
       <groupId>Ab</groupId>
       <version>0.0.4.3</version>
       <scope>system</scope>
-      <systemPath>C:/Users/av201/Desktop/workspace/Java/myChatBot/lib/Ab.jar</systemPath>
+    <systemPath>${basedir}/src/main/resources/lib/Ab.jar</systemPath>
     </dependency>
+    <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-web -->
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-web</artifactId>
+      <version>2.2.4.RELEASE</version>
+    </dependency>
+
   </dependencies>
 
   <build>
