+++ src/main/java/org/common/chatbot/engine/SimpleEngine.java
@@ -1,5 +1,8 @@
 package org.common.chatbot.engine;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public interface SimpleEngine {
 
     public String respondToMessage(String message);
