+++ src/main/resources/application.properties
@@ -1,0 +1,1 @@
+bot.telegram.token = 983716540:AAFGMtb1bmKJ8tNljKXctsumTQ4XYj607zo
